Add eslint Add prettier Set conventions
