+++ .prettierignore
@@ -1,0 +1,4 @@
+# Ignore artifacts:
+build
+coverage
+node_modules
