+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+  "singleQuote": false,
+  "trailingComma": "all",
+  "bracketSpacing": true
+}
