+++ src/App.css
@@ -1,53 +1,50 @@
 .top-bar {
-    filter: drop-shadow(0px 20px 10px rgba(0, 0, 0, 0.05));
-    background-color: #707070;
-    /*    make it flex in row direction*/
-
-
+  filter: drop-shadow(0px 20px 10px rgba(0, 0, 0, 0.05));
+  background-color: #707070;
+  /*    make it flex in row direction*/
 }
 
 .all-icons {
-    display: flex;
-    flex-direction: row; /* סידור בשורה */
-    justify-content: space-between; /* ריווח שווה בין הרכיבים */
-    align-items: center; /* יישור אנכי */
-    margin-right: 100px;
-    direction: rtl; /* כדי לתמוך בשפה מימין לשמאל */
-    gap: 10px; /* ריווח קבוע בין פריטים */
+  display: flex;
+  flex-direction: row; /* סידור בשורה */
+  justify-content: space-between; /* ריווח שווה בין הרכיבים */
+  align-items: center; /* יישור אנכי */
+  margin-right: 100px;
+  direction: rtl; /* כדי לתמוך בשפה מימין לשמאל */
+  gap: 10px; /* ריווח קבוע בין פריטים */
 }
 
 .top-bar-fonts {
-    font-size: 20px;
-    font-weight: 500;
-    line-height: 24px;
-    font-style: normal;
-    text-align: right;
-    color: #ffffff;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 24px;
+  font-style: normal;
+  text-align: right;
+  color: #ffffff;
 }
 
 .new-syllabus-btn {
-    text-decoration: none;
+  text-decoration: none;
 }
 
 .home-icon {
-    /*no background color*/
-    background-color: transparent;
-    border: none;
-    position: absolute;
-    width: 27px;
-    height: 24px;
-    margin-top: 15px;
-    margin-bottom: 13px;
+  /*no background color*/
+  background-color: transparent;
+  border: none;
+  position: absolute;
+  width: 27px;
+  height: 24px;
+  margin-top: 15px;
+  margin-bottom: 13px;
 }
 
 .person-icon {
-    background-color: transparent;
-    border: none;
-    position: absolute;
-    width: 25.1px;
-    height: 25.1px;
-    margin-right: 218px;
-    margin-top: 15px;
-    margin-bottom: 12px;
+  background-color: transparent;
+  border: none;
+  position: absolute;
+  width: 25.1px;
+  height: 25.1px;
+  margin-right: 218px;
+  margin-top: 15px;
+  margin-bottom: 12px;
 }
-
