--- src/ExemCard/ExamCard.module.css
@@ -1,140 +1,0 @@
-.examCard {
-    border-radius: 8px;
-    background: #fff;
-    box-shadow: 0 5px 3px rgba(0, 0, 0, 0.149);
-    display: flex;
-    margin-top: 13px;
-    width: 70%;
-    max-width: 100%;
-    padding: 0 0 0 36px;
-    gap: 40px 74px;
-    text-align: right;
-}
-
-@media (max-width: 991px) {
-    .examCard {
-        margin-right: 10px;
-        padding-left: 20px;
-    }
-}
-
-.statusIcon {
-    aspect-ratio: 1;
-    object-fit: contain;
-    object-position: center;
-    width: 34px;
-    align-self: start;
-    margin-top: 14px;
-}
-
-.statusBar {
-    background-color: rgba(255, 255, 255, 1);
-    display: flex;
-    /*width: 356px;*/
-    max-width: 100%;
-    height: 53px;
-}
-
-.statusContainer {
-    align-self: start;
-    display: flex;
-    margin-top: 6px;
-
-    gap: 40px 70px;
-    font-size: 14px;
-    color: rgba(255, 255, 255, 1);
-    font-weight: 700;
-    white-space: nowrap;
-    text-align: center;
-    letter-spacing: 0.47px;
-    line-height: 2;
-    flex-wrap: wrap;
-}
-
-@media (max-width: 991px) {
-    .statusContainer {
-        max-width: 100%;
-        white-space: initial;
-    }
-}
-
-.statusLabel {
-    border-radius: 10px;
-    background: #617CFF;
-    width: 78px;
-    height: 19px;
-    margin: auto ;
-    font-size: 14px;
-    font-family: 'Rubik', serif;
-    line-height: 22px;
-
-}
-
-@media (max-width: 991px) {
-    .statusLabel {
-        white-space: initial;
-    }
-}
-
-.visually-hidden {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    border: 0;
-}
-
-.facultyLabel {
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 1;
-    letter-spacing: 0.53px;
-    margin: auto 0;
-}
-
-.professorLabel {
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 1;
-    letter-spacing: 0.53px;
-    flex-grow: 1;
-    width: 113px;
-    margin: auto 0;
-}
-
-.courseSection {
-    display: flex;
-    gap: 18px;
-}
-
-.courseInfo {
-    display: flex;
-    flex-direction: column;
-    margin: auto 0;
-}
-
-.courseTitle {
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 1;
-    letter-spacing: 0.53px;
-}
-
-.examDate {
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 2;
-    letter-spacing: 0.13px;
-    align-self: end;
-    margin-top: 9px;
-}
-
-.statusIndicator {
-    border-radius: 0 8px 0 8px;
-    background-color: #617CFF;
-    width: 8px;
-    height: 80px;
-}
