--- src/Hello.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Hello() {
-    return <h2>שלום עולם!</h2>;
-}
-
-export default Hello;
