--- src/header/Header.module.css
@@ -1,113 +1,0 @@
-.headerContainer {
-    background-color: rgba(112, 112, 112, 1);
-    box-shadow: 0px 20px 10px rgba(0, 0, 0, 0.051);
-    display: flex;
-    height: 52px;
-    width: 100%;
-    gap: 20px;
-    /*flex-wrap: wrap;*/
-    justify-content: space-between;
-    padding: 11px 20px; /* צמצום הפדינג */
-    box-sizing: border-box; /* מונע חריגה מרוחב המסך */
-    direction: ltr;
-}
-
-@media (max-width: 991px) {
-    .headerContainer {
-        max-width: 100%;
-        padding: 0 20px;
-    }
-}
-
-.logoGroup {
-    align-self: start;
-    display: flex;
-    gap: 18px;
-    margin-left: 60px;
-}
-
-.logoImage {
-    aspect-ratio: 1;
-    object-fit: contain;
-    object-position: center;
-    width: 28px;
-}
-
-.secondaryLogo {
-    aspect-ratio: 1;
-    object-fit: contain;
-    object-position: center;
-    width: 26px;
-    align-self: start;
-}
-
-.userSection {
-    display: flex;
-    align-items: center;
-    gap: 20px; /* Adjust space between user info and notification icons */
-    justify-content: flex-end; /* Align content to the end for RTL */
-    width: 100%; /* Ensure it spans the full width */
-}
-
-
-.divider {
-    width: 0.5px;
-    height: auto;
-    background-color: rgba(183, 183, 183, 1);
-    margin-right: 30px;
-    margin-left: 30px;
-
-}
-
-.userInfo {
-    /*align-self: stretch;*/
-    display: flex;
-    align-items: center;
-    gap: 14px;
-    color: rgba(255, 255, 255, 1);
-    text-align: right;
-    letter-spacing: 0.19px;
-    /*margin: auto 0;*/
-    font: 500 20px/1.2 Rubik, sans-serif;
-}
-
-.userName {
-    flex-grow: 1;
-}
-
-.userIcon {
-    width: 30px; /* Adjust size for consistency */
-    height: 30px;
-    border-radius: 50%; /* Optional: circular icons */
-    object-fit: cover; /* Ensure the icon is fully visible */
-}
-
-.notificationSection {
-    align-self: stretch;
-    display: flex;
-}
-
-.notificationIcon {
-    width: 30px;
-    height: 30px;
-    object-fit: contain;
-}
-
-.homeButton {
-    background: none;
-    border: none;
-    margin-right: 60px;
-    cursor: pointer;
-}
-
-.visually-hidden {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    white-space: nowrap;
-    border: 0;
-}
