+++ src/pages/InstructorDetails.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "../styles/InstructorDetails.module.css";
 
-export const InstructorDetails = ({ email, name, officeHours, schedule }) => {
+export const InstructorDetails = ({ email, name, officeHours }) => {
   return (
     <section className={styles.instructorSection}>
       <h2 className={styles.sectionTitle}>פרטי המרצה</h2>
@@ -29,13 +29,13 @@ export const InstructorDetails = ({ email, name, officeHours, schedule }) => {
         <div className={styles.officeHours}>
           <h3 className={styles.subTitle}>שעות קבלה</h3>
           <div className={styles.scheduleInput}>
-            <div className={styles.checkbox} role="checkbox" tabIndex="0"/>
+            <div className={styles.checkbox} role="checkbox" tabIndex="0" />
 
             <input
-                type="text"
-                value={officeHours}
-                className={styles.input}
-                aria-label="שעות קבלה"
+              type="text"
+              value={officeHours}
+              className={styles.input}
+              aria-label="שעות קבלה"
             />
           </div>
         </div>
