+++ src/pages/LecturerDataEntryComplete.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { CourseDetailsPage } from "./CourseDetailsPage";
+
+function lecturerDataEntryComplete() {
+  return (
+    <div>
+      <CourseDetailsPage />
+      {/*<CourseHeader/>*/}
+      {/*<CourseContent/>*/}
+      {/*<InstructorDetails/>*/}
+      {/*<Bibliography/>*/}
+    </div>
+  );
+}
+
+export default lecturerDataEntryComplete;
