+++ src/styles/Bibliography.module.css
@@ -1,10 +1,8 @@
 .bibliography {
-
   background-color: #fff;
   border-radius: 5px;
   box-shadow: 0 20px 10px rgba(0, 0, 0, 0.051);
   padding: 18px 43px 92px;
-
 }
 
 .header {
@@ -49,7 +47,6 @@
   align-items: flex-start;
   gap: 20px;
   flex-direction: row-reverse;
-
 }
 
 .actions {
@@ -75,7 +72,7 @@
   gap: 13px;
   flex-grow: 1;
 }
-.citationItem{
+.citationItem {
   display: flex;
   flex-direction: row;
 }
