+++ src/styles/CourseHeader.module.css
@@ -1,108 +1,104 @@
 /* העיצוב הכללי של הכותרת */
 .header {
-    align-items: center;
-    display: flex;
-    flex-direction: column; /* כותרת למעלה, שאר המידע למטה */
-    background-color: #c9c9c9;
-    box-shadow: 0 20px 10px rgba(0, 0, 0, 0.051);
-    /*padding: 15px 0 8px;*/
-    /*width: 100%;*/
+  align-items: center;
+  display: flex;
+  flex-direction: column; /* כותרת למעלה, שאר המידע למטה */
+  background-color: #c9c9c9;
+  box-shadow: 0 20px 10px rgba(0, 0, 0, 0.051);
+  /*padding: 15px 0 8px;*/
+  /*width: 100%;*/
 }
 
 /* החלק העליון - כותרת */
 .titleSection {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    /*margin-bottom: 10px;*/
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  /*margin-bottom: 10px;*/
 }
 
 .titleText {
-    font-size: 18px;
-    font-weight: bold;
-    color: #555;
+  font-size: 18px;
+  font-weight: bold;
+  color: #555;
 }
 
 .titleIcon {
-    width: 16px;
-    height: 16px;
-    object-fit: contain;
+  width: 16px;
+  height: 16px;
+  object-fit: contain;
 }
 
 /* החלק התחתון - פרטי הקורס */
 .courseInfo {
-    display: flex;
-    justify-content: space-between; /* שמירה על פיזור בין הפרטים לסטטוס */
-    align-items: center;
-    border-top: 1px solid #a09f9f; /* קו מפריד */
-    padding-top: 10px;
+  display: flex;
+  justify-content: space-between; /* שמירה על פיזור בין הפרטים לסטטוס */
+  align-items: center;
+  border-top: 1px solid #a09f9f; /* קו מפריד */
+  padding-top: 10px;
 }
-
-
 
 /* פרטי הקורס */
 .courseDetails {
-    display: flex;
-    align-items: center;
-    gap: 10px; /* מרווחים בין הפריטים */
+  display: flex;
+  align-items: center;
+  gap: 10px; /* מרווחים בין הפריטים */
 }
 
 .facultyName,
 .courseName,
 .courseCode {
-    font: 500 20px/1.2 Rubik, sans-serif;
+  font:
+    500 20px/1.2 Rubik,
+    sans-serif;
 }
 
-
 .divider {
-    width: 2px;
-    height: 35px;
-    background-color: #a09f9f; /* צבע הקו המפריד */
+  width: 2px;
+  height: 35px;
+  background-color: #a09f9f; /* צבע הקו המפריד */
 }
 
 /* תגית סטטוס */
 .statusBadge {
-    background-color: #707070;
-    color: #fff;
-    border-radius: 15px;
-    padding: 5px 20px;
-    font-size: 14px;
-    font-weight: 500;
+  background-color: #707070;
+  color: #fff;
+  border-radius: 15px;
+  padding: 5px 20px;
+  font-size: 14px;
+  font-weight: 500;
 }
 
 .facultyInfo {
-    display: flex;
-    align-items: center;
-    gap: 8px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
 }
 
-
 .titleIcon {
-
-    width: 20px; /* גודל האייקון */
-    object-fit: contain; /* מבטיח שהאייקון יוצג במלואו */
+  width: 20px; /* גודל האייקון */
+  object-fit: contain; /* מבטיח שהאייקון יוצג במלואו */
 }
 
 .statusBadge {
-    background-color: #707070;
-    color: #fff;
-    border-radius: 30px;
-    padding: 3px 38px;
-    font-weight: 500;
-    letter-spacing: 0.83px;
+  background-color: #707070;
+  color: #fff;
+  border-radius: 30px;
+  padding: 3px 38px;
+  font-weight: 500;
+  letter-spacing: 0.83px;
 }
-
 
 /* התאמות למסכים קטנים */
 @media (max-width: 991px) {
-    .courseInfo {
-        flex-direction: column; /* מסדר את התוכן בטור */
-        align-items: flex-start;
-        gap: 10px; /* מוסיף מרווח בין פרטים לסטטוס */
-    }
+  .courseInfo {
+    flex-direction: column; /* מסדר את התוכן בטור */
+    align-items: flex-start;
+    gap: 10px; /* מוסיף מרווח בין פרטים לסטטוס */
+  }
 
-    .statusBadge {
-        align-self: flex-end; /* יישור הסטטוס לימין */
-        padding: 5px 15px;
-    }
+  .statusBadge {
+    align-self: flex-end; /* יישור הסטטוס לימין */
+    padding: 5px 15px;
+  }
 }
