+++ src/styles/InstructorDetails.module.css
@@ -54,7 +54,6 @@
   font-size: 14px;
   letter-spacing: 0.47px;
   color: #707070;
-
 }
 
 .officeHours {
